Refactor of mongoose init
