+++ bin/tilloo-cli-addjob.js
@@ -1,19 +1,8 @@
 #! /usr/bin/env node
 'use strict';
 
-const mongoose = require('mongoose');
 const commander = require('commander');
-
-const config = require('../lib/config');
 const jobs = require('../lib/jobs');
-
-mongoose.set('useNewUrlParser', true);
-mongoose.set('useFindAndModify', false);
-mongoose.set('useCreateIndex', true);
-mongoose.set('useUnifiedTopology', true);
-mongoose.connect(config.db);
-
-mongoose.Promise = global.Promise;
 
 function list(val) {
     return val.split(',');
