+++ bin/tilloo-cli-backup.js
@@ -1,18 +1,7 @@
 #! /usr/bin/env node
 'use strict';
 
-const mongoose = require('mongoose');
-
-const config = require('../lib/config');
 const Job = require('../models/job');
-
-mongoose.set('useNewUrlParser', true);
-mongoose.set('useFindAndModify', false);
-mongoose.set('useCreateIndex', true);
-mongoose.set('useUnifiedTopology', true);
-mongoose.Promise = global.Promise;
-mongoose.connect(config.db);
-
 
 Job.find({ deleted: false }, null, { sort: { name: 1 } }, function (err, jobs) {
     if (err) {
