+++ bin/tilloo-cli-listruns.js
@@ -1,20 +1,11 @@
 #! /usr/bin/env node
 'use strict';
 
-const mongoose = require('mongoose');
 const commander = require('commander');
 const Table = require('easy-table');
 const moment = require('moment');
 
-const config = require('../lib/config');
 const Run = require('../models/run');
-
-mongoose.set('useNewUrlParser', true);
-mongoose.set('useFindAndModify', false);
-mongoose.set('useCreateIndex', true);
-mongoose.set('useUnifiedTopology', true);
-mongoose.Promise = global.Promise;
-mongoose.connect(config.db);
 
 commander.version('0.0.1')
     .usage('<jobId>', 'Id of job')
@@ -24,7 +15,6 @@ if (commander.args.length !== 1) {
     commander.help();
     process.exit(1);
 }
-
 
 const table = new Table();
 Run.findRunsForJob(commander.args[0], { startedAt: 1 }, function (err, runs) {
