+++ bin/tilloo-cli-restore.js
@@ -2,22 +2,14 @@
 'use strict';
 
 const fs = require('fs');
-const mongoose = require('mongoose');
-const ObjectId = require('mongoose').Types.ObjectId;
+const mongoose = require('../lib/mongooseinit');
+const ObjectId = mongoose.Types.ObjectId;
 const commander = require('commander');
 const async = require('async');
 
 const config = require('../lib/config');
 const jobs = require('../lib/jobs');
 const Job = require('../models/job');
-
-mongoose.set('useNewUrlParser', true);
-mongoose.set('useFindAndModify', false);
-mongoose.set('useCreateIndex', true);
-mongoose.set('useUnifiedTopology', true);
-mongoose.Promise = global.Promise;
-mongoose.connect(config.db);
-
 
 commander.version('0.0.1')
     .usage('<file>')
