+++ lib/jobs.js
@@ -1,10 +1,11 @@
 'use strict';
 
-const ObjectId = require('mongoose').Types.ObjectId;
+const mongoose = require('../lib/mongooseinit');
+const ObjectId = mongoose.Types.ObjectId;
 const config = require('./config');
 const constants = require('./constants');
 const debug = require('debug')('tilloo:jobs');
-const rabbit = require('./rabbitFactory');
+const rabbit = require('./rabbitfactory');
 
 const Job = require('../models/job');
