+++ lib/k8s/eventstream.js
@@ -5,14 +5,14 @@
 const constants = require('../constants');
 
 const debug = require('debug')('tilloo:k8s/eventstream');
-const rabbit = require('../rabbitFactory');
+const rabbit = require('../rabbitfactory');
 
 const Checkpoint = require('../../models/checkpoint');
 const config = require('../config');
 const k8sJob = require('./job');
 const k8sClient = require('./clientFactory');
 
-console.info(`Eventstream connecting to ${config.rabbitmq.host}:${config.rabbitmq.port}`);
+debug(`publishing to ${config.rabbitmq.host}:${config.rabbitmq.port}`);
 
 const checkpointName = 'events';
