+++ lib/k8s/zombieruns.js
@@ -7,7 +7,7 @@ const constants = require('../constants');
 
 const async = require('async');
 const debug = require('debug')('tilloo:k8s/zombieruns');
-const rabbit = require('../rabbitFactory');
+const rabbit = require('../rabbitfactory');
 const moment = require('moment');
 
 const config = require('../config');
@@ -15,7 +15,7 @@ const Job = require('../../models/job');
 const k8sClient = require('./clientFactory');
 const Run = require('../../models/run');
 
-console.info(`ZombieRuns connecting to ${config.rabbitmq.host}:${config.rabbitmq.port}`);
+debug(`publishing to ${config.rabbitmq.host}:${config.rabbitmq.port}, queues: ${constants.QUEUES.STATUS}`);
 
 function updateStatus(message) {
     message.source = 'zombieruns';
