+++ lib/logforwarder.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const debug = require('debug')('tilloo:logforwarder');
-const mongoose = require('mongoose');
 const Tail = require('tail').Tail;
 const rabbit = require('./rabbitFactory');
 
@@ -9,14 +8,7 @@ const config = require('../lib/config');
 const constants = require('../lib/constants');
 const Log = require('../models/log');
 
-console.info(`Logforwarder connecting to ${config.rabbitmq.host}:${config.rabbitmq.port}, queue: ${constants.QUEUES.LOGGER}`);
-
-mongoose.set('useNewUrlParser', true);
-mongoose.set('useFindAndModify', false);
-mongoose.set('useCreateIndex', true);
-mongoose.set('useUnifiedTopology', true);
-mongoose.Promise = global.Promise;
-mongoose.connect(config.db);
+debug(`publishing to ${config.rabbitmq.host}:${config.rabbitmq.port}, queue: ${constants.QUEUES.LOGGER}`);
 
 function getRestartMaxDate(restart, runId, callback) {
     if (restart) {
