+++ lib/logwriter.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const rabbit = require('../lib/rabbitFactory');
+const rabbit = require('./rabbitfactory');
 
 const config = require('../lib/config');
 const constants = require('../lib/constants');
@@ -12,7 +12,6 @@ const debug = require('debug')('tilloo:logwriter');
 
 const logAppend = util.promisify(Log.append).bind(Log);
 
-console.info(`Listening to ${config.rabbitmq.host}:${config.rabbitmq.port} queue: ${constants.QUEUES.LOGGER}`);
 rabbit.subscribe(constants.QUEUES.LOGGER, async (message) => {
     debug('logger message', message);
     if (message && message.runId && message.output) {
