+++ lib/mongooseinit.js
@@ -1,0 +1,14 @@
+'use strict';
+
+const config = require('./config');
+
+const mongoose = require('mongoose');
+
+mongoose.set('useNewUrlParser', true);
+mongoose.set('useFindAndModify', false);
+mongoose.set('useCreateIndex', true);
+mongoose.set('useUnifiedTopology', true);
+mongoose.Promise = global.Promise;
+mongoose.connect(config.db);
+
+module.exports = mongoose;
