+++ lib/mqstatus.js
@@ -1,8 +1,9 @@
 'use strict';
 
-const ObjectId = require('mongoose').Types.ObjectId;
+const mongoose = require('../lib/mongooseinit');
+const ObjectId = mongoose.Types.ObjectId;
 
-const rabbit = require('./rabbitFactory');
+const rabbit = require('./rabbitfactory');
 
 const config = require('./config');
 const constants = require('./constants');
@@ -17,7 +18,6 @@ const runFindById = util.promisify(Run.findById).bind(Run);
 const runFindOneAndUpdate = util.promisify(Run.findOneAndUpdate).bind(Run);
 const jobFindByIdAndUpdate = util.promisify(Job.findByIdAndUpdate).bind(Job);
 
-console.info(`Listening to ${config.rabbitmq.host}:${config.rabbitmq.port} queue: ${constants.QUEUES.LOGGER}`);
 rabbit.subscribe(constants.QUEUES.STATUS, async (message) => {
     debug('status message', message);
     if (message && message.runId && message.status) {
