+++ lib/rabbitfactory.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const config = require('../lib/config');
-const constants = require('../lib/constants');
+const config = require('./config');
+const constants = require('./constants');
 const debug = require('debug')('tilloo:rabbitFactory');
 const rabbot = require('rabbot');
 
@@ -68,6 +68,7 @@ async function getInstance() {
 
 module.exports.subscribe = async function subscribe(queue, handler) {
     const r = await getInstance();
+    console.info(`Listening to ${config.rabbitmq.host}:${config.rabbitmq.port} queue: ${queue}`);
     r.handle(queue, async (message) => {
         debug(`queue: ${queue}, message: ${JSON.stringify(message)}`);
         if (message && message.body) {
