+++ lib/runs.js
@@ -7,10 +7,11 @@ const constants = require('./constants');
 const Run = require('../models/run');
 const Log = require('../models/log');
 const K8sJob = require('./k8s/job');
-const rabbit = require('./rabbitFactory');
+const rabbit = require('./rabbitfactory');
 
-console.info(`Runs connecting to ${config.rabbitmq.host}:${config.rabbitmq.port}`);
+
 const debug = require('debug')('tilloo:runs');
+debug(`publishing to ${config.rabbitmq.host}:${config.rabbitmq.port}, queue: ${constants.QUEUES.STATUS}`);
 
 const Runs = {};
