+++ models/run.js
@@ -1,7 +1,7 @@
 const constants = require('../lib/constants');
 
 const moment = require('moment');
-const mongoose = require('mongoose');
+const mongoose = require('../lib/mongooseinit');
 
 const Run = new mongoose.Schema({
     jobId: { type: mongoose.Schema.ObjectId, required: true, index: true },
@@ -9,18 +9,18 @@ const Run = new mongoose.Schema({
     path: { type: String, required: true },
     queueName: { type: String },
     timeout: { type: Number },
-    status: { type:String, enum: [constants.JOBSTATUS.BUSY, constants.JOBSTATUS.IDLE, constants.JOBSTATUS.FAIL, constants.JOBSTATUS.SUCCESS], default: constants.JOBSTATUS.IDLE },
-    result: { type:Number },
+    status: { type: String, enum: [constants.JOBSTATUS.BUSY, constants.JOBSTATUS.IDLE, constants.JOBSTATUS.FAIL, constants.JOBSTATUS.SUCCESS], default: constants.JOBSTATUS.IDLE },
+    result: { type: Number },
     createdAt: { type: Date, default: function () { return new Date(); }, index: true }, // eslint-disable-line brace-style
     updatedAt: { type: Date, default: function () { return new Date(); } }, // eslint-disable-line brace-style
-    startedAt:{ type: Date },
-    completedAt:{ type: Date },
-    pod: { type:String },
-    worker: { type:String }
+    startedAt: { type: Date },
+    completedAt: { type: Date },
+    pod: { type: String },
+    worker: { type: String }
 });
 
-Run.pre('save', function(done) {
-    if(!this.createdAt) {
+Run.pre('save', function (done) {
+    if (!this.createdAt) {
         this.createdAt = new Date();
     }
     this.updatedAt = new Date();
@@ -34,7 +34,7 @@ Run.statics.findRunsForJob = function findRunsForJob(jobId, sort, callback) {
 
 Run.statics.findRunsForJobPaginated = function findRunsForJobPaginated(jobId, page, pageSize, sort, callback) {
     Model.find({ jobId: new mongoose.Types.ObjectId(jobId) }, null, { sort: sort })
-        .skip((page-1) * pageSize)
+        .skip((page - 1) * pageSize)
         .limit(pageSize)
         .exec(callback);
 };
